+++ R/crew_package.R
@@ -19,13 +19,14 @@
 #'   [`rrq`](https://mrc-ide.github.io/rrq/),
 #'   [`clustermq`](https://mschubert.github.io/clustermq/),
 #'   and [`batchtools`](https://mllg.github.io/batchtools/).
+#' @importFrom data.table rbindlist
 #' @importFrom getip getip
 #' @importFrom mirai daemons mirai saisei server
 #' @importFrom nanonext mclock msleep random sha1 .unresolved
 #' @importFrom processx process
 #' @importFrom ps ps_handle ps_is_running ps_kill
 #' @importFrom R6 R6Class
-#' @importFrom rlang abort as_function enquo is_installed quo_squash
+#' @importFrom rlang abort as_function enquo is_installed is_named quo_squash
 #' @importFrom stats runif
 #' @importFrom tibble as_tibble new_tibble tibble
 #' @importFrom tidyselect all_of any_of contains ends_with eval_select